+++ src/trunk/apps/python/scevtstreams.py
@@ -37,9 +37,15 @@ class EventStreams(seiscomp3.Client.Application):
     self.commandline().addStringOption("Dump", "event,E", "event id")
     self.commandline().addIntOption("Dump", "margin,m", "time margin around the picked timewindow, default is 300")
     self.commandline().addStringOption("Dump", "streams,S", "comma separated list of streams per station to add, e.g. BH,SH,HH")
+    self.commandline().addOption("Dump", "all-streams", "dump all streams. If unused just streams with picks are dumped.")
     self.commandline().addIntOption("Dump", "all-components,C", "all components or just the picked one, default is True")
     self.commandline().addIntOption("Dump", "all-locations,L", "all components or just the picked one, default is True")
+    self.commandline().addOption("Dump", "all-stations", "dump all stations from the same network. If unused just stations with picks are dumped.")
+    self.commandline().addOption("Dump", "all-networks", "dump all networks. If unused just networks with picks are dumped. "\
+                                            "This option implies all-stations, all-locations, all-streams, all-components "\
+                                            "and will only provide the time window.")
     self.commandline().addOption("Dump", "resolve-wildcards,R", "if all components are used, use inventory to resolve stream components instead of using '?' (important when Arclink should be used)")
+    self.commandline().addOption("Dump", "caps", "dump in capstool format (Common Acquisition Protocol Server by gempa GmbH)")
     return True
 
   def validateParameters(self):
@@ -66,11 +72,27 @@ class EventStreams(seiscomp3.Client.Application):
       except: pass
 
       try:
+        self.allStreams = self.commandline().hasOption("all-streams")
+      except: pass
+
+      try:
         self.allComponents = self.commandline().optionInt("all-components") != 0
       except: pass
 
       try:
+        self.allStations = self.commandline().hasOption("all-stations")
+      except: pass
+
+      try:
+        self.allNetworks = self.commandline().hasOption("all-networks")
+      except: pass
+
+      try:
         self.allLocations = self.commandline().optionInt("all-locations") != 0
+      except: pass
+
+      try:
+        self.caps = self.commandline().hasOption("caps") != 0
       except: pass
 
       return True
@@ -125,19 +147,54 @@ class EventStreams(seiscomp3.Client.Application):
             streams = [rawStream + "?"]
 
         if self.allLocations == True:
-          loc = ""
+          if self.caps:
+            loc = "*"
+          else:
+            loc = ""
+
+        if self.allStations:
+          station = "*"
+        else:
+          station = pick.waveformID().stationCode()
+
+        if self.allNetworks:
+          net = "*"
+          station = "*"
+          loc = "*"
+        else:
+          net = pick.waveformID().networkCode()
 
         for s in streams:
-          line = minTime.toString("%F %T") + ";" + maxTime.toString("%F %T") + ";" \
-          + pick.waveformID().networkCode() + "." + pick.waveformID().stationCode() \
-          + "." + loc + "." +  s
+
+          if self.allStreams or self.allNetworks:
+            s = "*"
+
+          if self.caps:
+
+            line = minTime.toString("%Y,%m,%d,%H,%M,%S") + " " + maxTime.toString("%Y,%m,%d,%H,%M,%S") + " " \
+                   + net + " " + station \
+                   + " " + loc + " " +  s
+          else:
+            line = minTime.toString("%F %T") + ";" + maxTime.toString("%F %T") + ";" \
+                   + net + "." + station \
+                   + "." + loc + "." +  s
+
           lines.add(line)
 
         for s in self.streams:
           if s != rawStream:
-            line = minTime.toString("%F %T") + ";" + maxTime.toString("%F %T") + ";" \
-            + pick.waveformID().networkCode() + "." + pick.waveformID().stationCode() \
-            + "." + loc + "." +  s  + streams[0][2]
+
+            if self.allStreams or self.allNetworks:
+              s = "*"
+
+            if self.caps:
+
+              line = minTime.toString("%Y,%m,%d,%H,%M,%S") + " " + maxTime.toString("%Y,%m,%d,%H,%M,%S") + " " \
+                     + net + " " + station + " " + loc + " " +  s  + streams[0][2]
+            else:
+              line = minTime.toString("%F %T") + ";" + maxTime.toString("%F %T") + ";" \
+                     + net + "." + station + "." + loc + "." +  s  + streams[0][2]
+
             lines.add(line)
 
       for line in sorted(lines):